修改mapper接口1.0

+++ src/main/java/com/example/gxutrainbranch/entity/StudentMeasurePoint.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class StudentMeasurePoint {
     private int studentMeasurePointId;
-    private int activityId;
+    private long activityId;
     private double activityScore;
-    private int studentID;
+    private long  studentID;
 }
